% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdfiver.R
\name{check_md5_file}
\alias{check_md5_file}
\title{Check md5 checksum files}
\usage{
check_md5_file(file_loc, md5_file_loc = NULL, md5_hash = NULL)
}
\arguments{
\item{file_loc}{the file to create the md5 checksum file for}

\item{md5_file_loc}{the location of the md5 checksum file to check against, defaults to the input file with .md5 appended to the end if neither this or the md5_hash parameters are supplied}

\item{md5_hash}{the md5 hash to check the file against, takes precedense over md5_file_loc parameter}
}
\value{
True if the hashes match, False if the hashes do not match
}
\description{
Check md5 checksum against supplied file
}
\examples{
md5_matches <- check_md5_file('~/file_contents.tsv.gz');
md5_matches <- check_md5_file('~/file_contents.tsv.gz', md5_file_loc = '~/file_contents.tsv.gz.md5');
md5_matches <- check_md5_file('~/file_contents.tsv.gz', md5_hash = '7aedffa4687d37d4007bbd8e7fcf000d');
}
