% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdfiver.R
\name{check_sha256_file}
\alias{check_sha256_file}
\title{Check sha256 checksum files}
\usage{
check_sha256_file(file_loc, sha256_file_loc = NULL, sha256_hash = NULL)
}
\arguments{
\item{file_loc}{the file to create the sha256 checksum file for}

\item{sha256_file_loc}{the location of the sha256 checksum file to check against, defaults to the input file with .md5 (or .sha256 if option is used) appended to the end if neither this or the md5_hash parameters are supplied}

\item{sha256_hash}{the sha256 hash to check the file against, takes precedense over sha256_file_loc parameter}
}
\value{
True if the hashes match, False if the hashes do not match
}
\description{
Check sha256 checksum against supplied file
}
\examples{
sha256_matches <- check_sha256_file('~/file_contents.tsv.gz');
sha256_matches <- check_sha256_file('~/file_contents.tsv.gz', sha256_file_loc = '~/file_contents.tsv.gz.md5');
sha256_matches <- check_sha256_file('~/file_contents.tsv.gz', sha256_hash = '7aedffa4687d37d4007bbd8e7fcf000d');
}
