% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdfiver.R
\name{create_sha256_for_file}
\alias{create_sha256_for_file}
\title{Create sha256 checksum files}
\usage{
create_sha256_for_file(file_loc, sha256_file_loc = NULL)
}
\arguments{
\item{file_loc}{the file to create the sha256 checksum file for}

\item{sha256_file_loc}{the location of the md5 checksum file to create, defaults to the input file with .md5 appended to the end (or .sha256 if that option is used)}
}
\value{
The location of the sha256 checksum file
}
\description{
Create sha256 checksum files from supplied file
}
\examples{
sha256_from_file_loc <- create_md5_for_file('~/file_contents.tsv.gz');
sha256_from_file_loc <- create_md5_for_file('~/file_contents.tsv.gz', '~/file_contents.tsv.gz.sha256');
}
